% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgebuilder.R
\name{kg_builder}
\alias{kg_builder}
\title{kg_builder}
\usage{
kg_builder()
}
\description{
Creates an instance of a knowledge graph builder GUI within the
RStudio Viewer. Interactions on the viewer GUI client will be sent back to the
RStudio server for processing of operations.
}
