% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgplot.R
\docType{class}
\name{kgplot}
\alias{kgplot}
\title{kgplot}
\usage{
kgplot(name, nodes = NA, properties = NA)
}
\arguments{
\item{name}{Name of the knowledge graph}

\item{nodes}{Nodes existing in knowledge graph}

\item{properties}{Properties associated with knowledge graph}
}
\description{
Helper function for creating knowledge graph object
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the knowledge graph}

\item{\code{nodes}}{All the nodes that are in the knowledge graph}
}}

