% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{node}
\alias{node}
\title{node}
\usage{
node(name, in_edges, out_edges, properties)
}
\arguments{
\item{name}{Name of the node}

\item{in_edges}{in edges connected to node}

\item{out_edges}{out edges connected to node}

\item{properties}{properties associated with node}
}
\description{
Helper function for creating node object
}
